requirements.txt
fastapi
uvicorn
motor
python-dotenv
httpx
jinja2
python-telegram-bot
pytz
google-generativeai
